Finish workeffortEventsByPeriod api development

+++ src/main/java/com/kyle/microservices/beans/CalendarEntry.java
@@ -2,6 +2,8 @@ package com.kyle.microservices.beans;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 /**
  * Created by Kyle on 9/25/2017.
  */
@@ -9,6 +11,6 @@ public @Data
 class CalendarEntry {
     private Boolean startOfPeriod;
     private WorkEffort workEffort;
-    private Integer periodSpan;
+    private BigInteger periodSpan;
     private String calEntryRange;
 }
